Sensitive data processing for events https://github.com/AtlasOfLivingAustralia/extended-data-model/issues/60
These changes remove sensitive records from the elastic index and downloads. The following changes where made to achieve this:
Changes to verbatim- to-interpreted pipeline to apply inheritance at this stage (as opposed to index time) so that inheritance can be used to determine if record is sensitive
Changes to Elastic pipeline to remove sensitive records
Generation of a interpreted + verbatim Occurrence output with sensitive rules pre-applied for use in Spark SQL exports
Includes:

fixes for NPE for Inherited fields.
changes to accomodate updated ALA taxonomy in tests